+++ inc/mailcollector.class.php
@@ -719,6 +719,7 @@ class MailCollector  extends CommonDBTM {
 
                   // Followup case
                   $ticket = new Ticket();
+                  $ticketExist = $ticket->getFromDB($tkt['tickets_id']);
                   $fup = new ITILFollowup();
 
                   $fup_input = $tkt;
@@ -726,7 +727,29 @@ class MailCollector  extends CommonDBTM {
                   $fup_input['items_id'] = $fup_input['tickets_id'];
                   unset($fup_input['tickets_id']);
 
-                  if (!$ticket->getFromDB($tkt['tickets_id'])) {
+                  if ($ticketExist && Entity::getUsedConfig(
+                        'suppliers_as_private',
+                        $ticket->fields['entities_id']
+                     )) {
+                     // Get suppliers matching the from email
+                     $suppliers = Supplier::getSuppliersByEmail(
+                        $rejinput['from']
+                     );
+
+                     foreach ($suppliers as $supplier) {
+                        // If the supplier is assigned to this ticket then
+                        // the followup must be private
+                        if ($ticket->isSupplier(
+                              CommonITILActor::ASSIGN,
+                              $supplier['id'])
+                           ) {
+                           $fup_input['is_private'] = true;
+                           break;
+                        }
+                     }
+                  }
+
+                  if (!$ticketExist) {
                      $error++;
                      $rejinput['reason'] = NotImportedEmail::FAILED_OPERATION;
                      $rejected->add($rejinput);