+++ install/update_94_95.php
@@ -113,6 +113,21 @@ function update94to95() {
       'number'
    ], 'softwares_id_expire_number');
 
+   /** Private supplier followup in glpi_entities */
+   if (!$DB->fieldExists('glpi_entities', 'suppliers_as_private')) {
+      $migration->addField(
+         "glpi_entities",
+         "suppliers_as_private",
+         "integer",
+         [
+            'value'     => -2,               // Inherit as default value
+            'update'    => 0,                // Not enabled for root entity
+            'condition' => 'WHERE `id` = 0'
+         ]
+      );
+   }
+   /** /Private supplier followup in glpi_entities */
+
    // ************ Keep it at the end **************
    $migration->executeMigration();